revert dev:cache to rake task, fixes #23410
