+++ railties/lib/rails/tasks.rb
@@ -3,6 +3,7 @@ require 'rake'
 # Load Rails Rakefile extensions
 %w(
   annotations
+  dev
   framework
   initializers
   log
